Discourage NativeAllocationRegistries without size
We previously encouraged NativeAllocationRegistries that managed
small amounts of malloc'ed memory to be constructed without an explicit
size, since the size is no longer used by the GC.

But it is desirable to have a size for ahat and similar tools.
So this encourages framework developers to instead supply at least a
wild guess, so that such allocations once again show up in heap dumps.

Change the size-less creation function to implicitly specify
a guess of 98 bytes (about 100, but less likely to be used as a real
size parameter), so that something will show up in heap dumps.

Bug: 223225532
Test: Treehugger
Change-Id: I7a1ce5e646c1f0431dfebaa5ed53fae57f0bca8e